requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-chroma
langchain-groq
langchain-huggingface
chromadb
huggingface_hub
sentence-transformers
pypdf
python-dotenv
langchain_core
